qcacld-3.0: Add check for max mcs rssi index
Currently when host driver tries to give the max vht rate, it
accesses the rssi_mcs_tbl in a loop without checking for the
max index for rssi_mcs_tbl array. This may result in array out
of bound access.

To address above issue, add a check in the for loop for the
max array index check.

Change-Id: I9ab2306f5ed4be745127d8fa35154902eb394e66
CRs-Fixed: 3618765