qcacld-3.0: Update VDEV's legacy OSIF pointer on MAC update
Legacy OSIF pointer in VDEV points to the link info element
that the VDEV is currently assigned to in adapter. As part of
link switch the VDEV will be reassigned to the link info structure
that corresponds to the new IEEE link id and here the OSIF
pointer is properly updated in the VDEV.

However post disconnect all the VDEVs will be reordered with the
link info structure on which it is initially created and in this
reset logic of the legacy OSIF pointer is updated for only one
VDEV and leaving the next VDEV to point to different link info
than where it is currently assigned.

In the reset logic, update all the valid VDEV's OSIF pointer to
reflect the link info it is currently mapped to.

Change-Id: Iba541ee99c008fb610abdd086ae40164dde0e74f
CRs-Fixed: 3626410