qcacld-3.0: Return success if chan width is not set for set wifi config
Currently wifi set configuration vendor command returns error
in case if chan width attribute is not in the vendor command,
this results in complete failure of the vendor command and does
not give chance to parse and execute other attributes and also
this command always returns failure for all the attributes
if chan width is not present.

To address this issue, return success if chan width is not
present in the set wifi configuration vendor command.

Change-Id: Ifd8a857ae6aadd1aa2f8227507e1e27b477e1200
CRs-Fixed: 3628393