qcacld-3.0: Do not fill TX flags for RX data and mgmt pkts in pkt capture
Currently, add_rtap_ext flag which is used to update the tx_flags in
radiotap header is set for both TX and RX data/mgmt packets which
is wrong.

This change fixes the code to not set the add_rtap_ext flag for RX data
and management packets so that tx_flags is not filled in radiotap header
of RX data and management packets.

Change-Id: I9c5a28c9bd47c2567af5ddb3e8e9c95463fe710f
CRs-Fixed: 3639248