qcacld-3.0: Move invalid pmkid check to LIM
In current design, host validates the PMKID in the
assoc request of an SAE STA after the Peer creation
in CSR. For cases with invalid PMKID, the assoc response
with status INVALID_PMKID and the peer delete are queued
back-to-back to the firmware. In some cases, this assoc
response gets flushed out due to the peer delete cmd and
leads to discard of the assoc response frame in SAP.

To fix this, validate the PMKID before the peer creation
as part of the assoc request sanity checks.

Change-Id: I5aa71188e4482063c5a5173b9aebd54ac9891c17
CRs-Fixed: 3631253