qcacld-3.0: Enable beacon protection after SSR
Currently, host driver sets beacon protection flag in the vdev
during SAP bring up and send it to the target with beacon template.
But while SSR, driver destroys vdev and hence beacon protection
flag is reset and after SSR, driver does not set the beacon
protection flag. Thus, target does not enable beacon protection
and does not send MIC IE in the beacon frame.

So, to fix this, set the beacon protection flag in the vdev while
restarting the SAP.

Change-Id: I14f64463ed0d2e66a305160dcc7c810012dde2fb
CRs-Fixed: 3636385