qcacld-3.0: initiate soc recovery cleanup up on detecting link down
There could be a delay in detection of link down and the driver
recovery getting triggered, this can result in the different
raceconditions between the userspace and driver, one such case
is that userspace issued IFF_DOWN when the driver is trying to
send the scan done notification.

The scandone is dropped in the driver resulting in the corresponding
scan getting rejected.

Fix is to initiate soc recovery cleanup freeing up pending active
commands, completing the pending scan requests in the driver as soon
as the driver detects link down.

CRs-Fixed: 3648073
Change-Id: Ib3091ce9062dfada730617dfcd697e20c34a85cd