qca-wifi: Do not enable RCAC when not in RCAC DFS domain
RCAC should be enabled only for DFS domains FCC and MKK and not for
other DFS domains. However, the sanity check is not present and allows
RCAC to be enabled even for ETSI domain.

To fix this, add the domain check before enabling rcac. Also, in
"dfs_is_agile_rcac_enabled" function rcac domain check is modularized.

Change-Id: I350361e993da171c67261ffbbe6deefd241f7b67