qcacld-3.0: Update the BS state if userspace changes the bearer
Currently host driver does not update the state of the bearer
switch state machine when userspace changes the bearer, which
results in out of sync state of the bearer currently in use.

To address above issue, update the state of the bearer switch
state machine if user space changes the bearer.

Change-Id: Id71cfdf320a11665707a367b13ea932155af0413
CRs-Fixed: 3649793