qcacld-3.0: Decrement NDI session in failure cases also
Currently, host driver sends NDP END ALL commands to the target
and target responses with NDP_HOST_UPDATE and NDP_END_IND. But
due to some reason if host driver does not receive NDP_END_IND
then there is timeout in driver and it starts the NDP cleanup.
While NDP cleanup, driver does not decrement the NDP session
in the PCL. Although driver destroy NDI but there is entry
remaining in the PCL. Hence, it rejects the NAN Enable command.

So, to fix this, while NDP cleanup, decrement the NDP sessions
and change the NDI state to disconnected in the failure cases
also and increase the timeout for NDP_END_ALL request.

Change-Id: I3c5f8f49adced528cb7199831e4d4c13a690dde3
CRs-Fixed: 3648611