qcacld-3.0: Allow authentication frame from existing peer STA
When peer station sends authentication request and if peer was
not cleaned up earlier due to peer didn't send deauth or other
reason, then the deletion of that peer STA is initiated at lim
and deauthentication is sent to the STA to cleanup the peer.
This is taken care in lim_process_auth_frame_type1() API.
But the duplicate peer detection logic in the same API tries to
validate if ML peer is found on any vdev, and since the peer
already exists it always returns true.

This causes the auth from the peer to be responded with auth
failure and when peer tries to deauth, the unencrypted
deauth is dropped, resulting in peer not able to connect again.

So check if the ML peer for the peer that requests for
authentication is present on other vdev alone.

Refactor the peer cleanup logic into a new API.

Change-Id: If54eea72b9dd2634a7dd0b8e2c4a4c6b7261b93c
CRs-Fixed: 3650632