qcacld-3.0: Fill MLO mgr with roam info using self link addr
Host clears existing connection info in MLO mgr and fills
roamed AP info in the same order received in roam_sync.
This overrides the existing VDEV order only in MLO manager but
it can be in different order in HDD.

During link switch MAC addr update, HDD first updates the VDEV
to the new link addr and later updates MLO manager and this
changes the order in MLO mgr, this can lead to having misaligned
link-VDEV mapping.

As FW is expected to derive same link addr as what host derives
and uses the same during connection with roamed AP, so use link
addr from roam sync params as a key to search MLO manager to
update the new roamed AP info. This will not alter the VDEV order
in MLO manager and will align with HDD order.

Change-Id: Ib1048f90a7e1b64c4135da2474579cffb669a49d
CRs-Fixed: 3655528