qcacld-3.0: Send auth rsp with reason code 0xe for auth req with invalid seq num
In current design, SAP will ignore the peer STA's Authentication
request frame with wrong sequence number, and no Authentication
response frame will send to this peer STA in this case.

The correct action for this case on SAP mode is that sending to
Authentication response frame to this peer STA with Status Code 0xe
which means Received an Authentication frame with authentication
transaction sequence number out of expected sequence.

Change-Id: I431e9d6b64da4cfc785e19df3b608487a2251f10
CRs-Fixed: 3665549