qcacld-3.0: Get band information from firmware for beacon report logging
In the api lim_send_radio_measure_report_action_frame(),
the band info is fetched from the pe session for beacon report
response connectivity logging as a result it is observed that
the beacon report response frame sent over the air is on a
different band that is getting logged as part of
the advanced connectivity logging for MLO usecase.

Fetch the band information from the firmware as part of
wmi_mgmt_tx_compl_event_fixed_param to get the actual band on
which the frame was sent over the air in order to maintain
proper band info for connectivity logging.

Change-Id: I8a9631c0f81ca71feab21c19218d07088a1ed69f
CRs-Fixed: 3668913