qca-wifi-oss: Indicate CAC completion after puncturing CAC expiry
When we enable DFS-Puncturing and trigger test radar on any of the
subchannels, the radar hit channel is added to NOL. The AP continues to
operate in the same operating channel and BW by puncturing the radar hit
channel.

After NOL expiry, DFS module starts the CAC timer, and does CAC on the
NOL channel. If no radar is detected during this punctured CAC period,
the channel is unpunctured and the AP operates in the full BW.

Issue:
if we check the state of the channel after unpuncturing, the radar
sub-channel's state still shown as DFS_CAC_REQUIRED / PRECAC_REQUIRED.
This is incorrect as the channel has completed CAC after NOL expiry.

Solution:
To fix this, deliver CAC completion event to userspace after punctured
CAC expiry and cache the PRECAC_DONE status of the sub-channel when
operating in ETSI domain.

Change-Id: I261843aa2f2121cb37d67e5bd6a5179d89fd0b98