qcacld-3.0: Add timer to handle bearer switch request to wlan
Currently host driver sends the bearer switch request to switch
the bearer to wlan as soon as it receives this request. With this
implementation, there is a possibility that host driver might send
back to back bearer switch requests to switch the bearer to wlan
and non-wlan back to back. To avoid these back to back requests,
add an opportunistic timer in host driver which waits for specified
timeout before sending the bearer switch request to user space from
non-wlan to wlan.
Once the timer expires, then only send the bearer switch request
to user space to switch from non-wlan to wlan.
Host driver can not use this opportunistic timer logic for wlan
to non-wlan bearer switch requests as it needs to switch the bearer
to non-wlan to perform off-channel activities.

Change-Id: I68369b6870265819739af040017250ce6b6c6dfc
CRs-Fixed: 3692183