qcacld-3.0: Add bearer switch related changes
With this change, added below bearer switch related changes:
1. Cache bearer switch request to switch the bearer to non-wlan
in non-wlan state.
2. Send timeout response to FW for FW requested bearer switch
requests.
3. Do not cache bearer switch requests multiple times also do
not increment the ref count multiple times for the same
bearer switch request_id.
For example, if a bearer switch to wlan is already requested and
before response to this request is received, a non-wlan bearer
switch request is received, then that request is cached and state
machine waits for the ongoing request completion, once the ongoing
request to switch the bearer to wlan is completed then this cached
request to non-wlan is executed, so in that case there is no need
to cache this request again.
4. Do not cache bearer switch requests multiple times also do
not decreament the ref count multiple times for the same
bearer switch request_id.
For example if bearer switch to wlan is received in non-wlan
requested state, then the ref count is decreamented and request is
cached. Once the response to the current non-wlan bearer switch
request is received then switch to wlan will be executed, so in
that case there is not need to decrement the ref count and  to
cache the request again.
5. Send the response as success to FW if host does not forward the
BS request to user space.
For example, if bearer is already non-wlan and FW sends a request
to switch the bearer to non-wlan. host does not send this request
to user space but today it does not send the response to FW as well
which results in timeout at FW because of this. Add a logic to send
success response whenever host does not sends the request to user
space.

Change-Id: I141403b6f833877693c2137ecd935d886ecf4a48
CRs-Fixed: 3692445