qcacld-3.0: Retain the ll_lt_sap policy
Currently when usecase stops for the ll_lt_sap, user space
updates the ll_lt_sap policy to none and host driver also
updates this policy to none. In the host driver this ap
policy is used to check if the interface is ll_lt_sap or
normal SAP and based on this host driver takes various
decisions on concurrency scenario. If this policy is reset
then host treats this interface as normal SAP interface and
implements force SCC logic if the policy is not reset then
host driver treats this interface as l_lt_sap and implements
force mcc logic.
To avoid force scc logic for ll_lt_sap, do not update the ap
policy of ll_lt_sap interface.

Change-Id: If03025e4b4b0db45d7d4032de1906b0690c023ab
CRs-Fixed: 3693651