component_dev: Bitmap support for dcs_sim
dcs_sim command for OBSS intf previously supports only
single input argument. It is now modified to
support two input arguments. The first argument is 0x10
which represents the OBSS parameter. The second argument
is the bitmap of the interference.
This support is added only to the OBSS Interference
paramter of the dcs_sim command.

CRs-Fixed: 3696149
Change-Id: I4d05490e49e60481ed79789513d72a99dc8342df