qcacld-3.0: Correct repetitive conditional statement
Currently, while sending auth and action frames to kernel, driver
tries to find the adapter from destination address if destination
address is not broadcast. But driver checks for destination
address again as broadcast address for certain action frames
after adapter is found and returns from the function and it drops
the frame.

So, to fix this, remove the repetitive check for destination
address as it is checked at the beginning of function.

Change-Id: I21de38835989dc85a9bdbb3b6a019939a6e237f2
CRs-Fixed: 3700884