qcacld-3.0: Skip 6 GHz freq in SAP mandatory freq list
Currently, in host driver, policy manager prepares the mandatory
frequency list for SAP. If SAP is not able to bring up with
provided frequency due to some reason (such as ACS HW mode is
not compatible with provided frequency), then driver selects the
frequency from the mandatory frequency list.
But if SAP is not allowed for 6 GHz band (because ACS frequency
list does not have 6 GHz frequencies) and SAP mandatory frequency
list have 6 GHz frequencies and driver selects the 6 GHz frequency
to bring up the SAP, then SAP start fails.

So, to fix this, skip the adding the 6 GHz frequencies in the SAP
mandatory channel list if SAP is not allowed in 6 GHz band.

Change-Id: If5d1e864e2b22a89e43a6a4bbdcb11934b39d2ca
CRs-Fixed: 3697125