qca-wifi: Allow vdev start if the vdev is in MVR pending vdev list
Consider a scenario where a Repeater STA vap is connected to the RootAP
and it triggered internal CSA when Repater STA's channel is different
from the RootAP.
As part of ieee80211_synced_channel_switch timer function, Repeater STA
vap sets the MVR bit and moves to START_RESTART_PROGRESS state. This
function also dispatches FW_VDEV_RESTART event to the Repeater AP vaps.
During the issue scenario, one of the Repeater AP vap was in
START_START_PROGRESS state. FW_VDEV_RESTART event is not handled in this
state. From START_START_PROGRESS state Host does not send vdev start
command to FW for the AP vap as MVR is in progress. Therefore, the AP
vap is stuck in START_START_PROGRESS and leads to MVR timeout.

To fix the issue, allow sending vdev start command to FW if the MVR is
in progress and if this vdev ID is set in MVR peding vdev bitmap.
After sending the vdev up command to FW, host moves this AP vap to
START_RESTART_PROGRESS state as MVR is in progress.

Change-Id: I20109fc259d7216482dde4b0aa47518ce0ed45ac
CRs-Fixed: 3695825