qcacld-3.0: Reset sap mandatory channel list on stop_ap
Currently, sap mandatory channel list is not reset during
stop_ap. If selected band changes and gEnableSAPManadatoryChanList
is enabled then new band frequencies may be discarded via pcl
modification based on sap mandatory channels and sap may come
up on frequency which is not as per the selected band.

Fix is to reset sap mandatory channels when sap stops.
So that if sap starts again with different band previous
mandatory channels are not considered.

Change-Id: Ibb70acc921558ca48c6385950adf5e48a55d20ef
CRs-Fixed: 3708549