qcacld-3.0: Restrict ACTION_OUI_CONNECT_1X1 for DBS 2x2
Currently, Host driver is sending ACTION_OUI_CONNECT_1X1 to
firmware as a part of WMI_PDEV_CONFIG_VENDOR_OUI_ACTION_CMDID
for all the action OUI leading to 1x1 mode in case of 2x2
capability.

Current implementation is to send all the action OUI to the fw
and in case of 2x2 solution if matching OUI ACTION_OUI_CONNECT_1X1
is present connection will degrade to 1x1 connection.

This change is to restrict sending ACTION_OUI_CONNECT_1X1
in case of DBS 2x2 capability.

CRs-Fixed: 3711830
Change-Id: I8c5cbbbfe7cc2dc73d75121f92014d8bba22b263